Get rid of the commands/ subdir as it causes import problems
