+++ sysl2/sysl/types.go
@@ -1,6 +1,11 @@
 package main
 
-import sysl "github.com/anz-bank/sysl/src/proto"
+import (
+	sysl "github.com/anz-bank/sysl/src/proto"
+	"github.com/sirupsen/logrus"
+	"github.com/spf13/afero"
+	"gopkg.in/alecthomas/kingpin.v2"
+)
 
 type Visitor interface {
 	Visit(Element) error
@@ -84,4 +89,18 @@ type CmdContextParamDatagen struct {
 	modules     *string
 	plantuml    *string
 	classFormat *string
+}
+
+type ExecuteArgs struct {
+	Module        *sysl.Module
+	ModuleAppName string
+	Filesystem    afero.Fs
+	Logger        *logrus.Logger
+}
+
+type Command interface {
+	Configure(*kingpin.Application) *kingpin.CmdClause
+	Execute(ExecuteArgs) error
+	Name() string
+	RequireSyslModule() bool
 }
